requirements.txt
catboost==1.0.3
lightgbm==3.3.1
matplotlib==3.3.2
mlflow==1.25.0
numpy==1.19.5
pandas==1.4.1
pandas_profiling==3.1.0
phik==0.12.0
Pillow==9.1.0
plotly==4.14.3
scikit_learn==1.0.2
scipy==1.5.2
seaborn==0.11.2
streamlit==1.7.0
streamlit_pandas_profiling==0.1.3
sweetviz==2.1.3
xgboost==1.5.1
